# Avoid unneeded macros to cut down on generated code