--- a/lib/cublas/libcublas.jl
+++ b/lib/cublas/libcublas.jl
@@ -19,7 +19,9 @@ macro check(ex, errs...)
     end
 
     quote
-        res = @retry_reclaim err -> $check $(esc(ex))
+        res = retry_reclaim(err -> $check) do
+            $(esc(ex))
+        end
         if res != CUBLAS_STATUS_SUCCESS
             throw_api_error(res)
         end