--- a/lib/cudadrv/libcuda.jl
+++ b/lib/cudadrv/libcuda.jl
@@ -28,15 +28,13 @@ end
     end
 end
 
-macro check(ex)
-    quote
-        res = $(esc(ex))
-        if res != SUCCESS
-            throw_api_error(res)
-        end
-
-        nothing
+function check(f)
+    res = f()
+    if res != SUCCESS
+        throw_api_error(res)
     end
+
+    return
 end
 
 const CUdevice_v1 = Cint