--- a/lib/cudnn/src/base.jl
+++ b/lib/cudnn/src/base.jl
@@ -1,6 +1,8 @@
 function cudnnCreate()
     handle_ref = Ref{cudnnHandle_t}()
-    @check unsafe_cudnnCreate(handle_ref) CUDNN_STATUS_NOT_INITIALIZED CUDNN_STATUS_INTERNAL_ERROR
+    check(CUDNN_STATUS_NOT_INITIALIZED,CUDNN_STATUS_INTERNAL_ERROR) do
+      unsafe_cudnnCreate(handle_ref)
+    end
     return handle_ref[]
 end