--- a/lib/cudnn/src/libcudnn.jl
+++ b/lib/cudnn/src/libcudnn.jl
@@ -12,22 +12,16 @@ const cudaStream_t = CUstream
     end
 end
 
-macro check(ex, errs...)
-    check = :(isequal(err, CUDNN_STATUS_ALLOC_FAILED))
-    for err in errs
-        check = :($check || isequal(err, $(esc(err))))
+function check(f, errs...)
+    res = retry_reclaim(in((CUDNN_STATUS_ALLOC_FAILED, errs...))) do
+        f()
     end
 
-    quote
-        res = retry_reclaim(err -> $check) do
-            $(esc(ex))
-        end
-        if res != CUDNN_STATUS_SUCCESS
-            throw_api_error(res)
-        end
-
-        nothing
+    if res != CUDNN_STATUS_SUCCESS
+        throw_api_error(res)
     end
+
+    return
 end
 
 mutable struct cudnnContext end