--- a/lib/cufft/CUFFT.jl
+++ b/lib/cufft/CUFFT.jl
@@ -6,7 +6,7 @@ using ..CUDA_Runtime
 
 using ..CUDA
 using ..CUDA: CUstream, cuComplex, cuDoubleComplex, libraryPropertyType
-using ..CUDA: unsafe_free!, @retry_reclaim, initialize_context
+using ..CUDA: unsafe_free!, retry_reclaim, initialize_context
 
 using CEnum: @cenum