--- a/lib/curand/random.jl
+++ b/lib/curand/random.jl
@@ -49,7 +49,9 @@ function Random.seed!(rng::RNG, seed=make_seed(), offset=0)
     update_stream(rng)
     curandSetPseudoRandomGeneratorSeed(rng, seed)
     curandSetGeneratorOffset(rng, offset)
-    @check unsafe_curandGenerateSeeds(rng) CURAND_STATUS_PREEXISTING_FAILURE
+    check(CURAND_STATUS_PREEXISTING_FAILURE) do
+        unsafe_curandGenerateSeeds(rng)
+    end
     return
 end