--- a/lib/cusolver/CUSOLVER.jl
+++ b/lib/cusolver/CUSOLVER.jl
@@ -6,7 +6,7 @@ using ..CUDA_Runtime
 
 using ..CUDA
 using ..CUDA: CUstream, cuComplex, cuDoubleComplex, libraryPropertyType, cudaDataType
-using ..CUDA: @allowscalar, assertscalar, unsafe_free!, @retry_reclaim, initialize_context
+using ..CUDA: @allowscalar, assertscalar, unsafe_free!, retry_reclaim, initialize_context
 
 using ..CUBLAS: cublasFillMode_t, cublasOperation_t, cublasSideMode_t, cublasDiagType_t
 using ..CUSPARSE: cusparseMatDescr_t