--- a/lib/cusolver/dense.jl
+++ b/lib/cusolver/dense.jl
@@ -7,7 +7,9 @@ using ..CUBLAS: unsafe_batch
 
 function cusolverDnCreate()
   handle_ref = Ref{cusolverDnHandle_t}()
-  @check unsafe_cusolverDnCreate(handle_ref) CUSOLVER_STATUS_NOT_INITIALIZED CUSOLVER_STATUS_INTERNAL_ERROR
+  check(CUSOLVER_STATUS_NOT_INITIALIZED, CUSOLVER_STATUS_INTERNAL_ERROR) do
+    unsafe_cusolverDnCreate(handle_ref)
+  end
   return handle_ref[]
 end