--- a/lib/cusolver/sparse.jl
+++ b/lib/cusolver/sparse.jl
@@ -6,7 +6,9 @@ using ..CUSPARSE: CuSparseMatrixCSR, CuSparseMatrixCSC, CuMatrixDescriptor
 
 function cusolverSpCreate()
   handle_ref = Ref{cusolverSpHandle_t}()
-  @check unsafe_cusolverSpCreate(handle_ref) CUSOLVER_STATUS_NOT_INITIALIZED CUSOLVER_STATUS_INTERNAL_ERROR
+  check(CUSOLVER_STATUS_NOT_INITIALIZED, CUSOLVER_STATUS_INTERNAL_ERROR) do
+    unsafe_cusolverSpCreate(handle_ref)
+  end
   return handle_ref[]
 end