--- a/lib/cusparse/CUSPARSE.jl
+++ b/lib/cusparse/CUSPARSE.jl
@@ -6,7 +6,7 @@ using ..CUDA_Runtime
 
 using ..CUDA
 using ..CUDA: CUstream, cuComplex, cuDoubleComplex, libraryPropertyType, cudaDataType
-using ..CUDA: unsafe_free!, @retry_reclaim, initialize_context, i32, @allowscalar
+using ..CUDA: unsafe_free!, retry_reclaim, initialize_context, i32, @allowscalar
 
 using CEnum: @cenum