--- a/lib/custatevec/src/libcustatevec.jl
+++ b/lib/custatevec/src/libcustatevec.jl
@@ -15,22 +15,16 @@ const int2 = Tuple{Int32,Int32}
     end
 end
 
-macro check(ex, errs...)
-    check = :(isequal(err, CUSTATEVEC_STATUS_ALLOC_FAILED))
-    for err in errs
-        check = :($check || isequal(err, $(esc(err))))
+function check(f, errs...)
+    res = retry_reclaim(in((CUSTATEVEC_STATUS_ALLOC_FAILED, errs...))) do
+        f()
     end
 
-    quote
-        res = retry_reclaim(err -> $check) do
-            $(esc(ex))
-        end
-        if res != CUSTATEVEC_STATUS_SUCCESS
-            throw_api_error(res)
-        end
-
-        nothing
+    if res != CUSTATEVEC_STATUS_SUCCESS
+        throw_api_error(res)
     end
+
+    return
 end
 
 const custatevecIndex_t = Int64