--- a/lib/cutensor/src/libcutensor.jl
+++ b/lib/cutensor/src/libcutensor.jl
@@ -12,22 +12,16 @@ const cudaStream_t = CUstream
     end
 end
 
-macro check(ex, errs...)
-    check = :(isequal(err, CUTENSOR_STATUS_ALLOC_FAILED))
-    for err in errs
-        check = :($check || isequal(err, $(esc(err))))
+function check(f, errs...)
+    res = retry_reclaim(in((CUTENSOR_STATUS_ALLOC_FAILED, errs...))) do
+        f()
     end
 
-    quote
-        res = retry_reclaim(err -> $check) do
-            $(esc(ex))
-        end
-        if res != CUTENSOR_STATUS_SUCCESS
-            throw_api_error(res)
-        end
-
-        nothing
+    if res != CUTENSOR_STATUS_SUCCESS
+        throw_api_error(res)
     end
+
+    return
 end
 
 @cenum cutensorOperator_t::UInt32 begin