--- a/lib/cutensornet/src/libcutensornet.jl
+++ b/lib/cutensornet/src/libcutensornet.jl
@@ -12,22 +12,16 @@ const cudaStream_t = CUstream
     end
 end
 
-macro check(ex, errs...)
-    check = :(isequal(err, CUTENSORNET_STATUS_ALLOC_FAILED))
-    for err in errs
-        check = :($check || isequal(err, $(esc(err))))
+function check(f, errs...)
+    res = retry_reclaim(in((CUTENSORNET_STATUS_ALLOC_FAILED, errs...))) do
+        f()
     end
 
-    quote
-        res = retry_reclaim(err -> $check) do
-            $(esc(ex))
-        end
-        if res != CUTENSORNET_STATUS_SUCCESS
-            throw_api_error(res)
-        end
-
-        nothing
+    if res != CUTENSORNET_STATUS_SUCCESS
+        throw_api_error(res)
     end
+
+    return
 end
 
 @cenum cutensornetStatus_t::UInt32 begin