--- a/res/wrap/libcusparse_prologue.jl
+++ b/res/wrap/libcusparse_prologue.jl
@@ -10,20 +10,14 @@ const cudaStream_t = CUstream
     end
 end
 
-macro check(ex, errs...)
-    check = :(isequal(err, CUSPARSE_STATUS_ALLOC_FAILED))
-    for err in errs
-        check = :($check || isequal(err, $(esc(err))))
+function check(f, errs...)
+    res = retry_reclaim(in((CUSPARSE_STATUS_ALLOC_FAILED, errs...))) do
+        f()
     end
 
-    quote
-        res = retry_reclaim(err -> $check) do
-            $(esc(ex))
-        end
-        if res != CUSPARSE_STATUS_SUCCESS
-            throw_api_error(res)
-        end
-
-        nothing
+    if res != CUSPARSE_STATUS_SUCCESS
+        throw_api_error(res)
     end
+
+    return
 end