--- a/res/wrap/libcutensor_prologue.jl
+++ b/res/wrap/libcutensor_prologue.jl
@@ -17,7 +17,9 @@ macro check(ex, errs...)
     end
 
     quote
-        res = @retry_reclaim err->$check $(esc(ex))
+        res = retry_reclaim(err -> $check) do
+            $(esc(ex))
+        end
         if res != CUTENSOR_STATUS_SUCCESS
             throw_api_error(res)
         end