--- a/res/wrap/libnvml_prologue.jl
+++ b/res/wrap/libnvml_prologue.jl
@@ -18,13 +18,11 @@ function initialize_context()
     end
 end
 
-macro check(ex)
-    quote
-        res = $(esc(ex))
-        if res != NVML_SUCCESS
-            throw_api_error(res)
-        end
-
-        nothing
+function check(f)
+    res = f()
+    if res != NVML_SUCCESS
+        throw_api_error(res)
     end
+
+    return
 end