… lib/cudnn/test/tensor.jl
… res/wrap/README.md
… src/pool.jl
… test/apiutils.jl
… test/pool.jl