Apparently lowering is pretty slow, so expanding the many `@retry_reclaim` macros took significant time (around 6ms each). Replacing it with a regular function cuts down the time to around 1.5ms, significantly improving the time to compile the package.

Before:

```
julia> @time Base.compilecache(pkg)
 28.516434 seconds (4.16 k allocations: 379.914 KiB)

BenchmarkTools.Trial: 10000 samples with 998 evaluations.
 Range (min … max):  14.018 ns … 260.649 ns  ┊ GC (min … max): 0.00% … 87.61%
 Time  (median):     14.689 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   15.337 ns ±   4.029 ns  ┊ GC (mean ± σ):  0.40% ±  1.54%

      ▆█
  ▂▂▂▄██▇▃▃▂▂▂▃▆▅▄▄▃▄▄▃▃▄▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▁▁▂▁▁▁▁▁▁▁▁▁▂ ▃
  14 ns           Histogram: frequency by time         19.6 ns <

 Memory estimate: 16 bytes, allocs estimate: 1.
```

After:
```
julia> @time Base.compilecache(pkg)
 22.752074 seconds (4.16 k allocations: 379.914 KiB)

❯ jl +dev --project -e 'using CUDA, BenchmarkTools; display(@benchmark CUBLAS.cublasGetProperty(CUDA.MAJOR_VERSION))'
BenchmarkTools.Trial: 10000 samples with 998 evaluations.
 Range (min … max):  14.198 ns … 253.505 ns  ┊ GC (min … max): 0.00% … 88.39%
 Time  (median):     14.754 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   15.411 ns ±   4.154 ns  ┊ GC (mean ± σ):  0.42% ±  1.53%

      ▅█▂
  ▂▂▂▂███▄▃▂▂▂▂▂▃▅▆▅▃▄▄▃▃▃▃▃▃▃▃▃▃▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▁▁▁▁▂ ▃
  14.2 ns         Histogram: frequency by time         18.8 ns <

 Memory estimate: 16 bytes, allocs estimate: 1.
```

There's still about 12s of that spent in lowering, much of which spent expanding `@cuda`. But that is a separate issue.

Helps with https://github.com/JuliaGPU/CUDA.jl/issues/1684